Use esbuild to bundle the extension.
I hate the fact that i need a build system for an interpreted language.
I hate the fact that i need a build system for an interpreted language.
I hate the fact that i need a build system for an interpreted language.
I hate the fact that i need a build system for an interpreted language.
I hate the fact that i need a build system for an interpreted language.
I hate the fact that i need a build system for an interpreted language.

Though I understand why VSCode wants it.
See https://code.visualstudio.com/api/working-with-extensions/bundling-extension